Fix missing function
